Alby auth refactoring
