+++ src/app/auth/page.tsx
@@ -1,13 +1,7 @@
 'use client'
 
-import { fetchUserFromAlby } from '@/lib/alby'
 import { useSession } from 'next-auth/react'
 import { useEffect } from 'react'
-
-const userEndpoint = 'https://api.getalby.com/user/me'
-const tokenEndpoint = 'https://api.getalby.com/oauth/token'
-const clientId = process.env.NEXT_PUBLIC_ALBY_CLIENT_ID
-const clientSecret = process.env.NEXT_PUBLIC_ALBY_CLIENT_SECRET
 
 export default function Auth() {
   const { data: session, status } = useSession()
